get: shift to /snapshots route using express server
